Moves the markdown command to public commands package and updates import paths
