+++ example/cmd/simple/main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"github.com/matzefriedrich/cobra-extensions/example/commands"
 	"github.com/matzefriedrich/cobra-extensions/pkg/charmer"
+	builtin "github.com/matzefriedrich/cobra-extensions/pkg/commands"
 	"os"
 
 	"github.com/spf13/cobra"
@@ -12,6 +13,7 @@ func main() {
 
 	app := charmer.NewRootCommand("simple-example", "")
 
+	app.AddCommand(builtin.NewMarkdownDocsCommand(app))
 	app.AddCommand(commands.CreateHelloCommand())
 
 	AddCryptCommands(app)
